requirements.txt
asgiref==3.2.10
Django==3.0.5
django-filter==2.3.0
djangorestframework==3.11.1
djongo==1.3.3
flake8==3.8.3
importlib-metadata==1.7.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
zipp==3.1.0
